refactor: 주문 상품 반정규화 및 상품 통계 테이블 생성

+++ src/main/java/com/prgrms/nabmart/domain/item/service/response/FindItemDetailResponse.java
@@ -1,13 +1,13 @@
 package com.prgrms.nabmart.domain.item.service.response;
 
 public record FindItemDetailResponse(Long itemId, String name, int price, String description,
-                                     int quantity, double rate, int reviewCount, int discount,
-                                     int like, int maxBuyQuantity) {
+                                     int quantity, double rate, long reviewCount, int discount,
+                                     long like, int maxBuyQuantity) {
 
     public static FindItemDetailResponse of(final Long itemId, final String name, final int price,
         final String description,
         final int quantity,
-        final double rate, final int reviewCount, final int discount, final int like,
+        final double rate, final long reviewCount, final int discount, final long like,
         final int maxBuyQuantity) {
         return new FindItemDetailResponse(itemId, name, price, description, quantity, rate,
             reviewCount, discount, like, maxBuyQuantity);
